+++ CHANGELOG.rst
@@ -5,6 +5,9 @@
 Changelog
 =========
 
+* 1.4.0 [2018-02-05]: Shortened traceback of test methods to point only to the
+  line that uses the function that generates the exception
+
 * 1.3.1 [2018-01-18]: Minor package build fix
 
 * 1.3.0 [2018-01-18]: Dropped support for Python interpreter versions 2.6, 3.3